Replace Dashboard with Statistics; remove Transcribe Audio (#6)
* feat(statistics): add StatisticsView with three metric cards

* fix(statistics): route initial load through scheduleReload for unified cancellation

* chore(open-file): disable file-open routing to removed Transcribe Audio view

* chore(audio): mark AudioTranscribeView and AudioFileRow as orphaned with TODO

* feat(nav): replace Dashboard/Transcribe Audio with Statistics, default to History

* chore(metrics): delete dashboard-only files

* feat(statistics): restore hero banner and four metric cards

Brings back the 'You have saved X with VoiceInk' hero and the four
original counters (Sessions, Words Dictated, Words Per Minute, Keystrokes
Saved). Promotions, help/resources, and the copy-system-info footer are
still dropped.
